To make this work, we'll add more annotations to Battle. When you think of one resource relating to another - like how our Battle relates to the Programmer resource - there are 2 ways to express that relation. You can either link to it or you can embed it. Those are both just valid ways to think about expressing a link between 2 resources.
This @HATEOAS\Relation lets you do whichever you want. For a link, use the href. To embed something, use the embedded key and set it to an expression that points to which object you want to embed.

And actually, if you include both href and embedded, it'll create a link and embed it.

Before we run the test, add a 'And print last response', because I like to see how my endpoints look. Let's run it:

php vendor/bin/behat features/api/battle.feature:26
Awesome it passes! If you look - HATEOAS is doing all the work for us. We still have _links, but we also have _embedded. What's cool is that it goes out to the Programmer resource and serializes it. You end up with all the same properties as normal, and you even end up with its links. So a lot of things are falling into place accidentally.