Prize: New Behat Definition
And just like with links, since embedded data always lives under _embedded, I have a built-in definition you can choose to use if you want to.
Behind the scenes, this knows to look for all of this on the _embedded property.

And the test still passes. Now I'll take out the print last response. When it comes to linking and embedding, I hope you're feeling dangerous!