Let's go to one other endpoint - /api/programmers:
Put this into the Hal Browser:

http://localhost:8000/api/programmers
Here, things don't quite work out as well. The response body on the right is the same as the properties on the left. That might not look wrong to you, but the format for our collection is not correct yet according to HAL.

Hal Collection Format

If we look back at the HAL Primer doc, you'll see that this is what a collection resource should look like:

{
    '_links': {
        'self': {
            'href': 'http://example.org/api/user?page=3'
        },
        'first': {
            'href': 'http://example.org/api/user'
        },
        'prev': {
            'href': 'http://example.org/api/user?page=2'
        },
        'next': {
            'href': 'http://example.org/api/user?page=4'
        },
        'last': {
            'href': 'http://example.org/api/user?page=133'
        }
    }
    'count': 3,
    'total': 498,
    '_embedded': {
        'users': [
            {
                '_links': {
                    'self': {
                        'href': 'http://example.org/api/user/mwop'
                    }
                },
                'id': 'mwop',
                'name': 'Matthew Weier O'Phinney'
            },
            {
                '_links': {
                    'self': {
                        'href': 'http://example.org/api/user/mac_nibblet'
                    }
                },
                'id': 'mac_nibblet',
                'name': 'Antoine Hedgecock'
            }
        ]
    }
}
Collections are 'Collection Resources'

By the way, we've talked about resources like Programmer and Battle, but with the collection endpoints - like /api/programmers - the collection itself is known as a resource. So this is a collection resource. And in HAL, the way this works is that all of the items in the collection are considered to be embedded resources. If you imagine that this is /api/users, all the users live under the _embedded key. The only true properties - if you have any - relate to the collection itself: things like the total number of items in the collection or what page you're on.

And above, you'll see that this collection has pagination and it's using the links to help the client know how to get to other pages. We're going to do this exact same thing later, which is awesome and it'll be really easy.