Structuring Our Collections
But for now, I want to get things into this structure. First, before we implement it, you guys know what we're going to do, we're going to update our test. In our programmer.feature, the scenario for the collection was looking for a programmers property to be the array. But now, it's going to be _embedded.programmers. Let's go even one step further and say that the first item in this should be the UnitTester.

So its nickname should equal UnitTester. This is line 84, so let's run this test first to make sure it fails:

php vendor/bin/behat features/api/programmer.feature:85
And it does.