By using HAL, it gives our responses a nice, predictable structure. And while that's really awesome, I want to show you something that's even more awesome that we get for free by using Hal.
I'll google for Hal Browser. If you haven't seen this before, I'm not going to tell you want it is until we see it in action. I'll copy the Git URL:

git@github.com:mikekelly/hal-browser.git
Move into the web/ directory of the project and clone this into a browser/ directory:

cd web
git clone git@github.com:mikekelly/hal-browser.git browser
Great!

Opening up the Browser

Inside of this directory is a browser.html file - so let's surf to that instead of our actual project:

http://localhost:8000/browser/browser.html
And let me show you the directory - it's just that HTML file, some CSS and some JS.

What we get is the HAL Browser. This is an API client that knows all about dealing with HAL responses. We can put in the URL. Let's guess that there's a battle whose id is 1 in the database, and make a real request to that URL:

Put this into the Hal Browser:

http://localhost:8000/api/battles/1

We're actually getting a 404 right now, because apparently there's no battle with id 1.