I'll re-run a test that puts battles into the database so we have something to play with:
php vendor/bin/behat features/api/battle.feature:26

Properties, Links and Embedded

Perfect, this time it finds a battle. On the right you see the raw response headers and raw JSON. But on the left, it's actually parsing through those a little bit. It sees that these are the properties. So even though it has this big response body on the right, the properties are only the id and didProgrammerWin type of fields. It also parses out the links and puts them in a nice Links section. And lets us follow that link, which I'll do in a second. Down here, it also has the embedded resource. It shows us that we have an embedded programmer, so let's click that and then it does the same thing. It says: here are the properties, and even says, this embedded resource has a link to itself.

Let's follow the link to the programmer. This actually makes a new GET request to this URL - which is awesome - parses out the properties, and once again, we see the self link. So just by using HAL, we get this for free, which makes it really easy to learn your own API and you can even expose this if you want, to your end users on your site. If you want to try our API, just go to this URL and navigate your way around it.