Coding up the Programmer's Battles Endpoint
Let's get to work! Open ProgrammerController. We'll need a new route and I'll copy the 'show' route, since the URL will be really similar. We'll add the /battles in the end and change the method to listBattlesAction.

The route name isn't important yet, but we'll use it later to link. Let's call it api_programmers_battles_list.

Implementing this is going to be really easy! I'll put it right between showAction and listAction so I can steal from both. Ok, let's think about what we need to do. First, we need to find the Programmer for this nickname. We have code for this, so let's steal it.

If you find yourself repeating a lot of code like this, you can always create a private function inside your controller class and put it there. That's similar to what we've been doing by putting functions inside of BaseController.

The second thing we need to do is to find all of the battles that are linked to this programmer. I have a shortcut for this that I'll use.

The code might be different in your project, but this is just saying:

'Hey, go query the battle table where programmerId is equal to the id
of the programmer that we have.'
So what's cool is that from here, this is exactly like the listAction, because it's just a collection resource. So I'm going to grab everything from it, change the variable to , change the key to battles, and that's it!

So with almost no work, we'll run the test again, and it passes!

php vendor/bin/behat features/api/programmer.feature:95