Adding the battles Relation
Back on the Hal Browser, if we hit Go, we still don't have that link. We know that we can just add /battles onto the URL, but there's no link yet. Let's add it!

Open up the Programmer class and copy and paste to create a second Relation. This time the key will be battles, and below we'll grab the route name we created for the endpoint.

Then, everything else looks good, because this route does need the nickname. So you may want to write a test for this if it's really important, but I'm just going to go back to the Hal browser, click 'Go', and boom! We have a new battles link, which we can follow and see the collection resource. We can open up one of the battles, follow a link back to the related programmer and click to go back to the battles once again. We're surfing through our API, which is really cool!

Adding the Battle self Relation

If we click to look at a battle, you'll notice that we're missing one little thing. It has a programmer link, but no self link, which we really want every resource to have because it's a nice standard and it comes in handy. So let's go add this, which is really easy.