Other Link Attributes (like title)
When you look at the Links section, there are a few other columns like title, name/index and docs. One of the things that this is highlighting is that your links can have more than just that href. So in Homepage, let's give the link a title of 'The API Homepage'

Let's go back to the browser to see that title. And now we can give any link a little bit more information.

Hi, Welcome! Homepage Message

Let's keep going! Since this is the API homepage, you probably want to give the client a nice welcome message and maybe even link to the human-readable documentation. Even though this Homepage class isn't being pulled from the database doesn't mean that we can't have properties. Let's create a  property and set that to some hard-coded text. You could even put your documentation URL here.

And now, our API homepage is getting interesting! But the real purpose of this homepage is to have links to the actual resources that the API client will want. This is really easy as well. The most obvious resource the client may want is the programmers collection. So let's do this here. We'll say programmers and we'll link to /api/programmers. That route doesn't have a name yet, so let's call it api_programmers_list.
And now we can use it in the Relation. To be super friendly, we'll give this link a nice title as well.

So let's hit Go. Now we have a really great homepage. We can come here, we can see the message with a link to the documentation, we can visit all of the programmers, and now we're dangerous. From there we can follow links to a programmer, to that programmer's battles, and anything else our links lead us to.