Let's do some pagination! Technically, HAL doesn't have any opinion on how all of this should work. But fortunately, the rest of the REST world does. As you can see in this example, a great way to handle pagination is with links (from http://phlyrestfully.readthedocs.org/en/latest/halprimer.html#collections):
{
  '_links': {
    'self': {
      'href': 'http://example.org/api/user?page=3'
        },
      'first': {
        href': 'http://example.org/api/user'
      },
      'prev': {
        'href': 'http://example.org/api/user?page=2'
      },
      'next': {
        href: 'http://example.org/api/user?page=4'
      },
      'last': {
        'href': 'http://example.org/api/user?page=133'
      }
    }
    'count': 3,
    'total': 498,
    '_embedded': {
      'users': [
        {
          '_links': {
            'self': {
             'href': 'http://example.org/api/user/mwop'
           }
         },
            'id': 'mwop',
            'name': 'Matthew Weier O'Phinney'
            },
            {
              '_links': {
                'self': {
                  'href': 'http://example.org/api/user/mac_nibblet'
              }
            },
              'id': 'mac_nibblet',
              'name': 'Antoine Hedgecock'
            },
            {
              '_links': {
                'self': {
                  'href': 'http://example.org/api/user/spiffyjr'
                }
              },
              'id': 'spiffyjr',
              'name': 'Kyle Spraggs'
            }
        ]
    }
}
The keys here - first, prev, next and last - aren't accidental. These are Internet-wide standards like self. So if you use these names for your links, your API will be consistent with a lot of other APIs.

Pagination: Do it with Query Parameters

The other important thing to notice in this example is that pagination is done with query parameters. Technically, there are a lot of ways the client could tell us what page they want for a collection, like query parameters or request headers. But honestly, query parameters are the easiest way. In our case, we're going to follow what you see here exactly. And with the HATEOAS library, this will be easy.