Pagination Scenario
First, let's setup a scenario to test this in programmers.feature. In this scenario, we're going to do something really cool: we're going to follow the links for pagination. I want to be able to go to our collection resource, get the URL for this next link and make a second request to the next link and actually see what's on page 2.

For our pagination, we'll show 5 programmers per page. In the Given, we need to add a bunch of programmers to try this out - I'll paste in some very imaginative code that gives us 12 programmers in the database.

Everything after this will be very similar to the normal collection resource, so I'll grab a the second half of that scenario. I'll remove the status code 200, because we're already testing for this above. After I make the first GET request, we're going to parse through the response, find the next link and make a second GET request. I already have a built-in step definition to do exactly that. I'll just say: And I follow the 'next' link.

If you looked at the implementation behind this - which I wrote - it knows that we're using HAL, so it knows to look at _links, next, href and make a second GET request for that.

So on the second page, we'd expect there to be Programmer7, so we can say: And the '_embedded.programmers' property should contain 'Programmer7', because we expect that word to be somewhere inside that JSON. And we expect there to not be Programmer2 and for there to also not be Programmer11.

Programmer2 should be on page 1 and Programmer11 should be on page 3.

This starts on line 96, so let's try the scenario out:

php vendor/bin/behat features/api.programmer.feature:96
It fails of course because it can't find the next link - our collection response doesn't have that because we haven't done any of the work for it yet.