If we look at the HATEOAS documentation, they talk about pagination. To do pagination, you'll return this PaginatedRepresentation resource. So let's create that in ProgrammerController. So here is listAction where we're getting our programmers. Below this, we'll say {dollar_sign}paginated = new PaginatedRepresentation. My IDE just added the use statement at the top for us - make sure you have it.
This takes a number of different arguments. The first is the actual CollectionRepresentation. The second is the route name to the list endpoint, which for us is api_programmers_list. And it'll use this to generate the links like next, first and last. The third argument is any array of parameters that need to be passed to the route. So if the route had a nickname or id wildcard, you'd pass that here. But there aren't any wildcards in this route, so we'll pass an empty array. The next three arguments are the page we're on, the number of records we're showing per page, and the total number of pages. I'm going to invent a few variables and set them above in a moment.

Above, let's fill this in. Initially, I just want to get the next and last _links to show up, so I'm going to take some shortcuts and not worry about truly paginating the results quite yet. I'll hardcode these variables for now. I'll say that we're always on page 1, that the limit is always 5, and we can calculate the total number of pages. If we have 12 programmers, divided by 5 gives us 2.4, then we'll round that up with the ceil function.

Normally you'd use a library to help with pagination, but since I'm faking it, I'm just doing some manual work myself.

Finally, now that we have this {dollar_sign}paginated object, we'll pass it to createApiResponse.

Cool. We'll still returning all of the programmers, so I don't expect our test to pass, but let's try it:

php vendor/bin/behat features/api.programmer.feature:96
Our test fails, but it is getting further. Ah, it actually is following the next link and it is actually seeing that the Programmer7 resource is in the response. But then it's failing because Programmer2 and every other programmer is still there.