Adding Real Pagination
Let's turn this into real pagination! Since the page and limit are being passed as query parameters, let's use those. In my application, when I need request information, I just add a Request {dollar_sign}request argument to my controller.
Then for query parameters, I can say {dollar_sign}request->query->get('page') and the second argument is the default value if there is no page sent for some reasons. And the same for limit - we'll let the client control this, but we'll default to 5.

Then for query parameters, I can say {dollar_sign}request->query->get('page') and the second argument is the default value if there is no page sent for some reasons. And the same for limit - we'll let the client control this, but we'll default to 5.

Normally when you do pagination, you might do a LIMIT, OFFSET query to the database or pass which subset of records you want to some search engine, like Elastic Search. In both cases, we end up with just the 5 records we want, instead of all 10,000. Here, I'm going to be lazy and not do that. I'm just going to query for all of my programmers, then just use a little PHP array magic to only give us the ones we want. I'm just trying to keep things simple.

If I were doing this in a real project, I'd probably use a library called Pagerfanta(https://github.com/whiteoctober/Pagerfanta). It helps you paginate and has built-in adapters already for Doctrine, and Elastic Search. You can give it things like, we're on page 2, we're showing 5 results per page, and then it'll do the work to find only the results we need.

Instead of that, I'm going to paste in some manual logic and use array_slice.

So I'm querying for all of the programmers and then slicing those down to the ones I want. It's not an efficient way to do this, but we should have a functional, paginated endpoint now. So let's try the test again, and it works!

php vendor/bin/behat features/api.programmer.feature:96
Yes!

And to enjoy it a little bit more, we can go back to the Hal Browser, hit Go, and since we're on page 2, we see Programmers 6-10. We can follow links to the first page, then over to the last page to see only those 2 programmers. Now pagination isn't hard, and it'll be really consistent across your API.